refactor file_util
