+++ documentsearcher.vcxproj.filters
@@ -15,7 +15,7 @@
     <Filter Include="src\ui">
       <UniqueIdentifier>{5431964c-2dd6-41b5-9568-cacabfd61033}</UniqueIdentifier>
     </Filter>
-    <Filter Include="src\file_util">
+    <Filter Include="src\util">
       <UniqueIdentifier>{ac0087c5-68d0-4c9a-8f57-79169e3e91fd}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
@@ -44,11 +44,11 @@
     <ClCompile Include="src\search_ui.cpp">
       <Filter>src</Filter>
     </ClCompile>
-    <ClCompile Include="src\file_util\file_util.cpp">
-      <Filter>src\file_util</Filter>
-    </ClCompile>
     <ClCompile Include="src\index_ui.cpp">
       <Filter>src</Filter>
+    </ClCompile>
+    <ClCompile Include="src\util\file_util.cpp">
+      <Filter>src\util</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -73,11 +73,11 @@
     <ClInclude Include="src\search_ui.h">
       <Filter>src</Filter>
     </ClInclude>
-    <ClInclude Include="src\file_util\file_util.h">
-      <Filter>src\file_util</Filter>
-    </ClInclude>
     <ClInclude Include="src\index_ui.h">
       <Filter>src</Filter>
+    </ClInclude>
+    <ClInclude Include="src\util\file_util.h">
+      <Filter>src\util</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
