+++ src/index_ui.cpp
@@ -6,7 +6,7 @@
 #include <nlohmann/json.hpp>
 
 #include "app.h"
-#include "file_util/file_util.h"
+#include "util/file_util.h"
 #include "logger.h"
 
 using json = nlohmann::json;
