+++ src/lucene_api/index.cpp
@@ -1,7 +1,7 @@
 #include "api.h"
 #include "index.h"
 
-#include "../file_util/file_util.h"
+#include "../util/file_util.h"
